+++ clickhouse_backend/backend/operations.py
@@ -1,4 +1,6 @@
+from django.apps import apps
 from django.db.backends.base.operations import BaseDatabaseOperations
+from django.utils.functional import cached_property
 
 from clickhouse_backend import compat
 from clickhouse_backend.driver import JSON
@@ -278,10 +280,29 @@ class DatabaseOperations(BaseDatabaseOperations):
         else:
             return "match(%s, concat('(?i)', %s))"
 
+    @cached_property
+    def table_model_dict(self):
+        return {model._meta.db_table: model for model in apps.get_models()}
+
+    def _get_on_cluster(self, style, table, sync=False):
+        if table in self.table_model_dict:
+            cluster = getattr(self.table_model_dict[table], "cluster", None)
+            if cluster:
+                if sync:
+                    return (
+                        f"ON CLUSTER {style.SQL_FIELD(self.quote_name(cluster))} SYNC"
+                    )
+                return f"ON CLUSTER {style.SQL_FIELD(self.quote_name(cluster))}"
+        return ""
+
     def sql_flush(self, style, tables, *, reset_sequences=False, allow_cascade=False):
         return [
-            "%s %s"
-            % (style.SQL_KEYWORD("TRUNCATE"), style.SQL_FIELD(self.quote_name(table)))
+            "%s %s %s"
+            % (
+                style.SQL_KEYWORD("TRUNCATE"),
+                style.SQL_FIELD(self.quote_name(table)),
+                self._get_on_cluster(style, table),
+            )
             for table in tables
         ]